+++ metrics-core/src/main/java/com/yammer/metrics/core/VirtualMachineMetrics.java
@@ -77,8 +77,7 @@ public class VirtualMachineMetrics {
      * @return Heap initial JVM memory in bytes.
      */
     public double heapInit() {
-        return getMemoryMXBean().getHeapMemoryUsage().getInit() +
-                getMemoryMXBean().getNonHeapMemoryUsage().getInit();
+        return getMemoryMXBean().getHeapMemoryUsage().getInit();
     }
     /**
      * Returns the heap memory currently used by the current JVM.
@@ -86,8 +85,7 @@ public class VirtualMachineMetrics {
      * @return Heap memory currently used by JVM in bytes.
      */
     public double heapUsed() {
-        return getMemoryMXBean().getHeapMemoryUsage().getUsed() +
-                getMemoryMXBean().getNonHeapMemoryUsage().getUsed();
+        return getMemoryMXBean().getHeapMemoryUsage().getUsed();
     }
     /**
      * Returns the heap memory currently used by the current JVM.
@@ -95,8 +93,7 @@ public class VirtualMachineMetrics {
      * @return Heap memory currently used by JVM in bytes.
      */
     public double heapMax() {
-        return getMemoryMXBean().getHeapMemoryUsage().getMax() +
-                getMemoryMXBean().getNonHeapMemoryUsage().getMax();
+        return getMemoryMXBean().getHeapMemoryUsage().getMax();
     }
     /**
      * Returns the heap memory committed to the JVM.
@@ -104,8 +101,7 @@ public class VirtualMachineMetrics {
      * @return Heap memory currently committed to the JVM in bytes.
      */
     public double heapCommitted() {
-        return getMemoryMXBean().getHeapMemoryUsage().getCommitted() +
-                getMemoryMXBean().getNonHeapMemoryUsage().getCommitted();
+        return getMemoryMXBean().getHeapMemoryUsage().getCommitted();
     }
 
     /**